# services/control-plane/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RunIt API
  description: |
    API for managing projects, runs, secrets, and deployments in RunIt.

    ## Authentication
    All endpoints (except health checks and metrics) require a valid Bearer token in the Authorization header.
    Use an API key or JWT token.

    ## Rate Limiting
    - Authenticated users: 120 requests per minute
    - Anonymous users: 60 requests per minute
  version: 0.1.0
  contact:
    name: RunIt
    url: https://runit.dev
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.runit.dev
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  schemas:
...
    Project:
      type: object
      required:
        - project_id
        - project_slug
        - name
        - status
        - created_at
        - updated_at
      properties:
        project_id:
          type: string
          format: uuid
          description: Unique project identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        project_slug:
          type: string
          description: URL-friendly project slug
          example: "my-api-project-a1b2c3d4"
        name:
          type: string
          description: Project display name
          example: "My API Project"
        owner_id:
          type: string
          description: User ID of project owner
          example: "user_123abc"
        status:
          type: string
          enum: [draft, deploying, live, failed]
          description: Current project deployment status
          example: "live"
        deployed_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of last successful deployment
        deploy_error:
          type: string
          nullable: true
          description: Error message from failed deployment
        runtime_url:
          type: string
          format: uri
          nullable: true
          description: URL where the deployed project is accessible
          example: "https://my-project.modal.run"
        detected_env_vars:
          type: array
          items:
            type: string
          description: Environment variables detected in the code
          example: ["API_KEY", "DATABASE_URL"]
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/Endpoint'
          description: List of API endpoints extracted from the project
        versions:
          type: array
          items:
            $ref: '#/components/schemas/Version'
          description: List of project versions
        created_at:
          type: string
          format: date-time
          description: Project creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Project last update timestamp
    Version:
      type: object
      required:
        - version_id
        - version_hash
        - created_at
        - status
      properties:
        version_id:
          type: string
          format: uuid
          description: Unique version identifier
        version_hash:
          type: string
          description: Short SHA256 hash of the code bundle
          example: "a1b2c3d4e5f6"
        created_at:
          type: string
          format: date-time
          description: Version creation timestamp
        status:
          type: string
          enum: [building, ready, failed]
          description: Version build status
    Endpoint:
      type: object
      required:
        - id
        - method
        - path
      properties:
        id:
          type: string
          description: Endpoint identifier (path-based)
          example: "GET_/users"
        endpoint_id:
          type: string
          description: Endpoint identifier (alias for id)
          example: "GET_/users"
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
          description: HTTP method
          example: "GET"
        path:
          type: string
          description: URL path pattern
          example: "/users/{id}"
        summary:
          type: string
          description: Brief description of the endpoint
          example: "Get user by ID"
        description:
          type: string
          description: Detailed description of the endpoint
...
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: New project display name
        slug:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{1,58}[a-z0-9]$"
          description: New URL-friendly project slug (3-60 chars, lowercase alphanumeric + hyphens)
...
    Run:
      type: object
      required:
        - run_id
        - project_id
        - version_id
        - endpoint_id
        - status
        - created_at
      properties:
        run_id:
          type: string
          format: uuid
          description: Unique run identifier
        project_id:
          type: string
          format: uuid
          description: Project ID
        version_id:
          type: string
          format: uuid
          description: Version ID used for the run
        endpoint_id:
          type: string
          description: Endpoint that was executed
        status:
          type: string
          enum: [queued, running, success, error, timeout]
          description: Current run status
        created_at:
          type: string
          format: date-time
          description: Run creation timestamp
        started_at:
          type: string
          format: date-time
          description: Run start timestamp
        completed_at:
          type: string
          format: date-time
          description: Run completion timestamp
        duration_ms:
          type: integer
          description: Execution duration in milliseconds
        created_by:
          type: string
          description: User ID who created the run
        result:
          $ref: '#/components/schemas/RunResult'
    RunResult:
      type: object
      properties:
        http_status:
          type: integer
          description: HTTP status code returned by the endpoint
          example: 200
        content_type:
          type: string
          description: Response content type
          example: "application/json"
        json:
          type: object
          description: Response body (if JSON)
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/Artifact'
          description: Generated artifacts (files, images, etc.)
        redactions_applied:
          type: boolean
          description: Whether PII redactions were applied
        error_class:
          type: string
          description: Error class (if failed)
          example: "ValueError"
        error_message:
          type: string
          description: Error message (if failed)
        suggested_fix:
          type: string
          description: Suggested fix for the error
        logs:
          type: string
          description: Execution logs
    Artifact:
      type: object
      required:
        - name
        - size
        - mime_type
        - download_url
      properties:
        name:
          type: string
          description: Artifact filename
        size:
          type: integer
          description: File size in bytes
        mime_type:
          type: string
          description: MIME type
          example: "image/png"
        download_url:
          type: string
          format: uri
          description: URL to download the artifact
    CreateRunRequest:
      type: object
      required:
        - project_id
        - version_id
        - endpoint_id
      properties:
        project_id:
          type: string
          format: uuid
          description: Project ID to run
        version_id:
          type: string
          format: uuid
          description: Version ID to run
        endpoint_id:
          type: string
          description: Endpoint ID to execute
        params:
          type: object
          additionalProperties: true
          description: URL/path parameters
        json:
          type: object
          additionalProperties: true
          description: JSON request body
        headers:
          type: object
          additionalProperties:
            type: string
          description: Request headers
        files:
          type: array
          items:
            type: object
            required:
              - filename
              - data
              - content_type
            properties:
              filename:
                type: string
              data:
                type: string
                format: byte
                description: Base64-encoded file content
              content_type:
                type: string
          description: File uploads
        lane:
          type: string
          enum: [cpu, gpu]
          default: cpu
          description: Resource lane (CPU or GPU)
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 300
          default: 60
          description: Execution timeout in seconds
    CreateRunResponse:
      type: object
      required:
        - run_id
        - status
      properties:
        run_id:
          type: string
          format: uuid
          description: Created run ID
        status:
          type: string
          enum: [running]
          description: Initial run status

    # ==================== Secrets ====================
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Project not found"
        code:
          type: string
          description: Error code for programmatic handling
          example: "PROJECT_NOT_FOUND"
        details:
          type: string
          description: Additional error details
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Validation error message
          example: "Missing required fields: name, source_type"
        received:
          type: string
          description: Value that was received (for debugging)
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          examples:
            missingFields:
              value:
                error: "Missing required fields: name, source_type"
            invalidFormat:
              value:
                error: "Invalid secret key format. Use UPPERCASE_SNAKE_CASE (A-Z, 0-9, _)"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
    Forbidden:
      description: Not authorized to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not authorized"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            projectNotFound:
              value:
                error: "Project not found"
            runNotFound:
              value:
                error: "Run not found"
            versionNotFound:
              value:
                error: "Version not found"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded. Try again in 60 seconds."
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per minute
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in the current window
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
paths:
...
  /v1/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Project ID

    get:
      tags: [Projects]
      summary: Get project details
      description: Returns detailed information about a specific project including versions and endpoints
      operationId: getProject
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Projects]
      summary: Update a project
      description: Update project name and/or slug
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  project_id:
                    type: string
                    format: uuid
                  project_slug:
                    type: string
                  name:
                    type: string
                  owner_id:
                    type: string
                  status:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Slug already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Projects]
      summary: Delete a project
      description: Permanently deletes a project and all associated data (versions, secrets, runs, share links)
      operationId: deleteProject
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - project_id
                properties:
                  success:
                    type: boolean
                    example: true
                  project_id:
                    type: string
                    format: uuid
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== Project Endpoints ====================
...
  /v1/projects/{id}/runs:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Project ID

    get:
      tags: [Runs]
      summary: List runs for a project
      description: Returns recent runs for a project, ordered by creation time
      operationId: listProjectRuns
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of runs to return
      responses:
        '200':
          description: List of runs
          content:
            application/json:
              schema:
                type: object
                required:
                  - runs
                  - total
                properties:
                  runs:
                    type: array
                    items:
                      type: object
                      properties:
                        run_id:
                          type: string
                          format: uuid
                        endpoint_id:
                          type: string
                        status:
                          type: string
                          enum: [queued, running, success, error, timeout]
                        created_at:
                          type: string
                          format: date-time
                        duration_ms:
                          type: integer
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== Deploy ====================
...
  /v1/runs:
    post:
      tags: [Runs]
      summary: Create a new run
      description: |
        Executes an endpoint and returns a run ID for tracking.
        The execution happens asynchronously - poll GET /v1/runs/{id} for status.
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
            example:
              project_id: "550e8400-e29b-41d4-a716-446655440000"
              version_id: "660e8400-e29b-41d4-a716-446655440001"
              endpoint_id: "POST_/analyze"
              json:
                text: "Hello world"
              lane: "cpu"
              timeout_seconds: 60
      responses:
        '202':
          description: Run created and started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/runs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Run ID

    get:
      tags: [Runs]
      summary: Get run status and result
      description: Returns the current status and result (if completed) of a run
      operationId: getRun
      responses:
        '200':
          description: Run status and result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
              examples:
                running:
                  summary: Run in progress
                  value:
                    run_id: "770e8400-e29b-41d4-a716-446655440002"
                    project_id: "550e8400-e29b-41d4-a716-446655440000"
                    version_id: "660e8400-e29b-41d4-a716-446655440001"
                    endpoint_id: "POST_/analyze"
                    status: "running"
                    created_at: "2024-01-15T10:30:00Z"
                    started_at: "2024-01-15T10:30:01Z"
                success:
                  summary: Successful run
                  value:
                    run_id: "770e8400-e29b-41d4-a716-446655440002"
                    project_id: "550e8400-e29b-41d4-a716-446655440000"
                    version_id: "660e8400-e29b-41d4-a716-446655440001"
                    endpoint_id: "POST_/analyze"
                    status: "success"
                    created_at: "2024-01-15T10:30:00Z"
                    started_at: "2024-01-15T10:30:01Z"
                    completed_at: "2024-01-15T10:30:05Z"
                    duration_ms: 4200
                    result:
                      http_status: 200
                      content_type: "application/json"
                      json:
                        sentiment: "positive"
                        confidence: 0.95
                error:
                  summary: Failed run
                  value:
                    run_id: "770e8400-e29b-41d4-a716-446655440002"
                    project_id: "550e8400-e29b-41d4-a716-446655440000"
                    version_id: "660e8400-e29b-41d4-a716-446655440001"
                    endpoint_id: "POST_/analyze"
                    status: "error"
                    created_at: "2024-01-15T10:30:00Z"
                    started_at: "2024-01-15T10:30:01Z"
                    completed_at: "2024-01-15T10:30:02Z"
                    duration_ms: 1500
                    result:
                      http_status: 0
                      content_type: "application/json"
                      error_class: "ValueError"
                      error_message: "Invalid input: text cannot be empty"
                      suggested_fix: "Provide a non-empty 'text' field in the request body"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== Billing ====================
...